Subcontexts and Handler class
Now subcontexts can be selected. I hope the design is good enough, it
was the only "easy" way I could think of!

I created a Handler class inside the web_application project. We need to
decide whether this will be the best idea or not.

The interface BingAPI/Search with the GUI is working.