Implementation of Refiner class; Update on the temp Database
The Refiner class is now implemented, ths being called from the Handler
class. The database connector is now a static class.

I tried my best to avoid this, but in order to set up the temp database
(and be able to run the application) its necessary to run some SQL
scripts. Please, read the file  READ_ME.docx, cause there I give
complete description how to do it. The scripts are in the same location
as the READ_ME file.