Result link fix; Back button fix; Project clean-up
Fixed the problem with the results loading an error page.
When the user clicked the browser back button, the page was cached, but
not entirely cached, so I decided to take out the cache feature.
Therefore, by hitting back button, a new page is loaded. This however is
just applicable for the default.aspx. For the results I did not see any
problem with cache.

Moreover, I delete some files that somehow were still inside the
folders, but not in the project. I also deleted the Operands class, as
we are not using it anymore.

The application should now work smoothly.